Refactor command package to use small caps
