+++ src/main/java/duke/command/UnmarkCommand.java
@@ -1,20 +1,20 @@
-package duke.Command;
+package duke.command;
 
 import duke.DukeException;
 import duke.Storage;
 import duke.TaskList;
 
-public class DeleteCommand extends Command {
+public class UnmarkCommand extends Command {
 
     private int index;
 
-    public DeleteCommand(String commandParams) {
+    public UnmarkCommand(String commandParams) {
         this.index = Integer.parseInt(commandParams);
     }
 
     @Override
     public String executeCommand(Storage storage, TaskList tasks) throws DukeException {
-        String result = tasks.deleteTaskWithResult(index);
+        String result = tasks.unMarkTaskWithResult(index);
         storage.save(tasks);
         return result;
     }
